Fix conflicts due to new dependencies

+++ src/main/java/de/embl/cba/bdv/utils/loaders/HDF5UnsignedLongAccessHack.java
@@ -38,9 +38,19 @@ import java.util.LinkedHashMap;
 import java.util.Map.Entry;
 
 import static bdv.img.hdf5.Util.reorder;
-import static ch.systemsx.cisd.hdf5.hdf5lib.H5D.*;
-import static ch.systemsx.cisd.hdf5.hdf5lib.H5S.*;
-import static ch.systemsx.cisd.hdf5.hdf5lib.HDF5Constants.*;
+import static hdf.hdf5lib.H5.H5Dclose;
+import static hdf.hdf5lib.H5.H5Dget_space;
+import static hdf.hdf5lib.H5.H5Dopen;
+import static hdf.hdf5lib.H5.H5Dread;
+import static hdf.hdf5lib.H5.H5Sclose;
+import static hdf.hdf5lib.H5.H5Screate_simple;
+import static hdf.hdf5lib.H5.H5Sget_simple_extent_dims;
+import static hdf.hdf5lib.H5.H5Sselect_hyperslab;
+import static hdf.hdf5lib.HDF5Constants.H5P_DEFAULT;
+import static hdf.hdf5lib.HDF5Constants.H5S_MAX_RANK;
+import static hdf.hdf5lib.HDF5Constants.H5S_SELECT_SET;
+import static hdf.hdf5lib.HDF5Constants.H5T_NATIVE_FLOAT;
+import static hdf.hdf5lib.HDF5Constants.H5T_NATIVE_ULONG;
 
 public class HDF5UnsignedLongAccessHack implements IHDF5UnsignedLongAccess
 {
@@ -59,9 +69,9 @@ public class HDF5UnsignedLongAccessHack implements IHDF5UnsignedLongAccess
 
 	private class OpenDataSet
 	{
-		final int dataSetId;
+		final long dataSetId;
 
-		final int fileSpaceId;
+		final long fileSpaceId;
 
 		public OpenDataSet( final String cellsPath )
 		{
@@ -176,7 +186,7 @@ public class HDF5UnsignedLongAccessHack implements IHDF5UnsignedLongAccess
 		reorder( min, reorderedMin );
 
 		final OpenDataSet dataset = openDataSetCache.getDataSet( new ViewLevelId( timepoint, setup, level ) );
-		final int memorySpaceId = H5Screate_simple( reorderedDimensions.length, reorderedDimensions, null );
+		final long memorySpaceId = H5Screate_simple( reorderedDimensions.length, reorderedDimensions, null );
 		H5Sselect_hyperslab( dataset.fileSpaceId, H5S_SELECT_SET, reorderedMin, null, reorderedDimensions, null );
 		H5Dread( dataset.dataSetId, H5T_NATIVE_ULONG, memorySpaceId, dataset.fileSpaceId, numericConversionXferPropertyListID, dataBlock );
 		H5Sclose( memorySpaceId );
@@ -201,7 +211,7 @@ public class HDF5UnsignedLongAccessHack implements IHDF5UnsignedLongAccess
 		reorder( min, reorderedMin );
 
 		final OpenDataSet dataset = openDataSetCache.getDataSet( new ViewLevelId( timepoint, setup, level ) );
-		final int memorySpaceId = H5Screate_simple( reorderedDimensions.length, reorderedDimensions, null );
+		final long memorySpaceId = H5Screate_simple( reorderedDimensions.length, reorderedDimensions, null );
 		H5Sselect_hyperslab( dataset.fileSpaceId, H5S_SELECT_SET, reorderedMin, null, reorderedDimensions, null );
 		H5Dread( dataset.dataSetId, H5T_NATIVE_FLOAT, memorySpaceId, dataset.fileSpaceId, numericConversionXferPropertyListID, dataBlock );
 		H5Sclose( memorySpaceId );
